Make overwriting IRC colours optional
The IRC colours are coded in the main HTML file. As I got it, it's not
intended that the IRC colours are customised / overwritten by a custom
style for consistency.
I prefer a uniform look and feel and prefer the IRC colours fit into the
style. But other people might see this different so they can disable this
behaviour by passing the --default-irc-colours options to the generator
script.